Local mongodb setup without data
