+++ Backend/controllers/resultController.js
@@ -1,8 +1,8 @@
-import { scrapeResults } from '../utils/scraper.js';
+// import { scrapeResults } from '../Scrapper/scraper.js';
 
 export const getResults = async (req, res) => {
     try {
-        const results = await scrapeResults();
+        // const results = await scrapeResults();
         res.status(200).json(results);
     } catch (error) {
         res.status(500).json({ error: 'Failed to fetch results' });
