+++ Backend/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "cors": "^2.8.5",
     "express": "^5.1.0",
+    "mongoose": "^8.15.1",
     "puppeteer": "^24.9.0"
   }
 }
