+++ Backend/server.js
@@ -3,6 +3,7 @@ import path from "path";
 import {dirname} from "path";
 import { fileURLToPath } from "url";
 import apiRoutes from "./routes/apiRoutes.js";
+import connectDB from "./config/mongodb.js";
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const app = express();
 
@@ -12,6 +13,7 @@ app.set("views", path.join(__dirname, "../Frontend/views"));
 // Serve static files like css, js, and images directly from here.
 app.use(express.static(path.join(__dirname, "../Frontend/public")));
 
+connectDB();
 // Route to render the home page
 app.use("/api", apiRoutes);
 app.get("/", (req, res) => {
