--- Backend/utils/results.js
@@ -1,8 +1,0 @@
-export const scrapeResults = async () => {
-    // Mock data for demonstration purposes
-    return [
-        { name: "Alice", rollNo: "101", cgpa: 3.8 },
-        { name: "Bob", rollNo: "102", cgpa: 3.6 },
-        { name: "Charlie", rollNo: "103", cgpa: 3.9 },
-    ];
-};
